+++ src/multiplesOfThreeAndFive/src/MultiplesOfThreeAndFive.java
@@ -1,3 +1,5 @@
+package multiplesOfThreeAndFive.src;
+
 import java.util.ArrayList;
 import java.util.List;
