+++ src/problemSeven/src/ProblemSeven.java
@@ -1,3 +1,5 @@
+package problemSeven.src;
+
 public class ProblemSeven {
 
     public static void main(String[] args) {
